requirements.txt
groq
sentence-transformers
qdrant-client
pymongo
langchain
langchain-community
langchain-text-splitters
llama_index
streamlit
python-dotenv
beautifulsoup4
trafilatura
playwright
requests
